requirements.txt
beautifulsoup4>=4.12.3
certifi>=2024.7.4
charset-normalizer>=3.3.2
click>=8.1.7
idna>=3.7
numpy>=2.0.1
pandas>=2.2.2
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
pytz>=2024.1
PyYAML>=6.0.1
requests>=2.32.3
six>=1.16.0
soupsieve>=2.5
tzdata>=2024.1
urllib3>=2.2.2
pytest>=7.0
